Add authntication and autorization for deleting a post

+++ controllers/api/v1/posts_api.js
@@ -21,16 +21,24 @@ module.exports.index = async function (req, res) {
 module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
-    post.remove();
 
-    await Comment.deleteMany({ post: req.params.id });
+    if (post.user == req.user.id) {
+      post.remove();
 
-    return res.status(200).json({
-      message: "Posts and assciated comments deleted successfully!",
-    });
-  } catch (error) {
-    return res.status(500).json({
-      message: "Internal server error",
+      await Comment.deleteMany({ post: req.params.id });
+
+      return res.json(200, {
+        message: "Post and associated comments deleted successfully!",
+      });
+    } else {
+      return res.json(401, {
+        message: "You cannot delete this post!",
+      });
+    }
+  } catch (err) {
+    console.log("********", err);
+    return res.json(500, {
+      message: "Internal Server Error",
     });
   }
 };
